Add Dialogflow CX Webhook Python Sample
This sample is Dialogflow agent setup Python script and a Cloud Function that
serves as a webhook for the same Dialogflow agent. The script creates a
Dialogflow agent that uses the Cloud Function as a webhook. The webhook sends
back a static response that is configured to send a response to the user
communicating with the Dialogflow agent.

The sample depends on the Dialogflow CX Python client library to make Dialogflow
CX API calls and gcloud to deploy the Cloud Function.